requirements.txt
beautifulsoup4==4.12.3
chromadb==0.5.5
Flask==3.0.3
flask_mail==0.10.0
flask_session==0.8.0
Flask_Testing==0.8.1
keras==3.5.0
langchain==0.2.14
langchain_chroma==0.1.2
langchain_community==0.2.12
langchain_openai==0.1.22
llama_index==0.10.65
mysql_connector_repackaged==0.3.1
nltk==3.9.1
numpy==2.1.0
python-dotenv==1.0.1
python_docx==1.1.2
Requests==2.32.3
sendgrid==6.11.0
sentence_transformers==3.0.1